refactor: Minor files/folders reallocation.
